main buttons created for UI
Simplified UI has two buttons on the main screen, a start/stop button and a clean/spin button. Both buttons are checked buttons, meaning that toggle their state when clicked. Start and stop starts and stops either the clean or spin mode. A status label is place at the bottom of the main screen.