Add custom sized gear icon for settings page
LVGL allows you to customize the pre-canned symbols available in lv_font library API. I created a 50x50 px gear icon that resided in the lower right corner of the display. The plan is to invoke a settings screen to set various speed and duration parameters on that screen. I also created a HOME icon and the BACK icon for future use.